+++ Silver4/11399.cpp
@@ -8,13 +8,13 @@ int main()
     ios::sync_with_stdio(false);
     cin.tie(0);
     
-    uint16_t n;
-    uint16_t* p;
+    uint32_t n;
+    uint32_t* p;
     uint64_t Answer = 0;
     uint64_t Temp = 0;
 
     cin >> n;
-    p = new uint16_t[n];
+    p = new uint32_t[n];
 
     for(int i = 0; i < n; ++i)
     {
@@ -23,7 +23,7 @@ int main()
 
     sort(p, p+n);
 
-    for(uint16_t i = 0; i < n; ++i)
+    for(uint32_t i = 0; i < n; ++i)
     {
         Temp = (Temp + p[i]);
         Answer += Temp;
